+++ NEWS.md
@@ -1,5 +1,6 @@
 # tweetrbot 0.0.1
 
+* Tweet the most retweeted of the month with `top_tweets()`
 * Stores tweets with a `#`
 * Retweet not already retweeted
 * Check if process is running